many Arduino and M5 libraries

+++ M5Unit-ACSSR/src/M5_ACSSR.cpp
@@ -7,7 +7,7 @@ bool M5_ACSSR::begin(TwoWire *wire, uint8_t sda, uint8_t scl, uint8_t addr) {
     _addr = addr;
     _sda  = sda;
     _scl  = scl;
-    _wire->begin(_sda, _scl, 100000UL);
+    _wire->begin((int)_sda, (int)_scl, 100000UL);
     delay(10);
     _wire->beginTransmission(_addr);
     uint8_t error = _wire->endTransmission();
